pyproject.toml
[project]
name = "podcastist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ctranslate2",
    "faster-whisper",
    "httpx",
    "jinja2",
    "mlx-whisper; sys_platform == 'darwin' and platform_machine == 'arm64'",
    "numpy",
    "onnxruntime",
    "openai-whisper",
    "opencc-python-reimplemented>=0.1.7",
    "requests",
    "torch",
    "torchaudio",
    "torchvision",
    "tqdm",
]
[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
# 這裡告訴 uv：如果在 Windows 且有 CUDA 需求，去哪裡抓 torch
torch = [
  { index = "pytorch-cu121", marker = "sys_platform == 'win32'" }
]
torchaudio = [
  { index = "pytorch-cu121", marker = "sys_platform == 'win32'" }
]
torchvision = [
  { index = "pytorch-cu121", marker = "sys_platform == 'win32'" }
]
